+++ src/contract/assembled_transaction.ts
@@ -422,6 +422,10 @@ export class AssembledTransaction<T> {
     this.built = this.raw.build();
     this.simulation = await this.server.simulateTransaction(this.built);
 
+    // need to force re-calculation of simulationData for new simulation
+    delete this.simulationResult;
+    delete this.simulationTransactionData;
+
     if (Api.isSimulationSuccess(this.simulation)) {
       this.built = assembleTransaction(
         this.built,